+++ packages/core/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./brand.js";
 export * from "./types.js";
+export * from "./flag-reasons.js";
 export * from "./git.js";
 export * from "./ingest.js";
 export * from "./parse.js";